requirements.txt
asgiref==3.5.2
backports.zoneinfo==0.2.1
certifi==2022.6.15
charset-normalizer==2.0.12
Django==4.0.5
django-cors-headers==3.13.0
django-js-asset==2.0.0
django-mptt==0.13.4
django-phonenumber-field==6.3.0
django-phonenumbers==1.0.1
djangorestframework==3.13.1
idna==3.3
mysqlclient==2.1.1
phonenumbers==8.12.50
PyJWT==2.4.0
pyotp==2.6.0
pytz==2022.1
requests==2.28.0
sqlparse==0.4.2
twilio==7.9.3
urllib3==1.26.9
